[rockchip_ebc] try to split overlapping areas into four subareas during refresh so that the non-overlapping parts can start to refresh as soon as possible and we only need to wait for the overlapping part.
The number of areas to split while preparing each frame can be limited.
I'm not sure if this is really required, but I fear that too many splits
could slow down the refresh thread.

Splitting areas can produce areas that do not align with full bytes (4
bit/byte), so we also try to account for odd start/end clips.