[rockchip_ebc] The current driver iteration does not guarantee consistency between the list of currently-worked on damaged areas (snapshot of ctx->queue taken at the beginning of each frame) and the framebuffer content (ctx->final). As such it is possible that the content of the framebuffer changes before a given area can be drawn, potentially leading to garbled screen content. This effects is hugely dependent on the nature of drawing calls emitted by individual applications. Large scheduled areas tend to be good, but if an application sends large bursts of overlapping/overwriting areas then bad things happen. The bug/effect is also triggered if area splitting is done to increase drawing performance.
For example, this can be nicely seen under Gnome when
chaotically moving the nautilus window.

This patch is not a fix but somewhat reduces the impact by moving the
splinlock guarding the ctx->queue so it guards both the whole
frame-prepartion phase of the partial refresh function and the
framebuffer blitting function.

An alternative that also greatly reduces the effect is to copy the whole
framebuffer before preparing a given frame. However, this has a huge
performance impact and thus is not feasible if we still want to to
real-time drawings.