[rockchip-ebc] Try to improve latency of the refresh thread
When measuring individual frame-processing times it became evident that
some frames took way longer than others (usually frame processing takes
about 15 ms, including waking up from the completion, but 30 ms
frame-processing time was sometimes observed). This is mainly caused by
the use of the spinlock, which encompasses large sections of code both
in the partial refresh function and the atomic_update function (I think
it is generally suggested to use mutexes in those cases - to be
investigated at another time).

The mitigation strategy implemented in this patch is as follows: In the
partial refresh thread, use spin_trylock to acquire the lock. If this
fails, no new clips will be processed in this frame, but frame execution
will be more consistent. To give the partial_refresh thread a better
chance to get the lock without delay, add some delays to the
atomic_update function AFTER releasing the lock there. This prevents
overly eager applications to swamp the update queue. Note that the
choice of delay time is the result of not very much testing. Too small
values will lead to not enough clips being processed, while too large
values will make everything slower.

My personal observation is that drawing in xournalpp gets somewhat
smoother due to reduced variations in drawing updates. The downside is
that this "optimization" leads to somewhat increased artifacting.

Lets see if we need to revert this patch....