[tps65185] when rebasing from v5.17-rc6-11023-gdadcf190596b to 6.2 the vdrive regulator, from time to time, failed to resume within the allowed time (ca. 83 ms), leading to a complete crash of the graphical stack and most of the kernel.
My understanding of the problem is that sometimes the vdrive activating
takes a bit longer, probably due to the active discharge feature of the
tps65185 chip (the REG_ENABLE register shows that the VEE_EN is 1 over
the whole waiting time, indicating that the chip is waiting for the
discharge before proceeding with the full activation of all rails).

Based on some log output the discharge time is a least 300 ms. Until
further investigation provides additional information, add an
off_on_delay to the vdrive description of 450ms. This implies that after
runtime suspend was triggered the ebc driver takes at least 450 ms to
reactivate.

Note that this behavior was not previously observed due to a bug in the
regulator code that effectively prevented the activation-timeout from
being triggered. See bug fix commit 8d8e16592022c9650df8aedfe6552ed478d7135b