drm/rockchip: ebc: delay return of plane_atomic_update function
We know that the hardware does not like to repeatedly update large areas
of the ebc display in short succession (ebc or memory issues?). Trying
to update often can lead to artifacts or strong ghosting.

The solution here would be to just limit the refresh rate, but due to
the current structure of the driver the ebc panel is reported to
GNOME/sway as having a refresh rate of 80 Hz, leading to quite a lot of
refreshes that have no effect other than introduce artifacts.

We here try to limit the impact by adding artificial delays to the
update function, which effectively rate limits the compositor. However,
we would like to retrain fast refreshes for small areas (e.g., when
drawing).

Two approaches have been investigated:

- Using the updated area as independent variable, compute a delay by
  using a polynomial of second order. This approach is commented in the
patch and did not lead to satisfactory results because it delays smaller
areas too much.
- The second approach just introduced a threshold area: A very small
  delay (2 ms) is used for areas below 100.000 pixels, and a large delay
(150 ms) is used for larger areas.

The second approach yielded surprisingly nice results: Fast drawing in
xournalpp, but only marginal ghosting when scrolling in
firefox/xournalpp. Also, it seems that general GNOME usage also
benefited from these delays.

The two delays can be controlled using the sysfs parameters delay_a
(small areas) and delay_b (large areas), both in [mu s]:

/sys/module/rockchip_ebc/parameters/delay_a
/sys/module/rockchip_ebc/parameters/delay_b