drm/rockchip: ebc: reduce artificial update delays to 1 ms
Previously, artificial delays were introduced in the panel_atomic_update
function to yield better interactions with the refresh thread, and, more
importantly, to reduce artifacting/ghosting for larger screen updates.
With the last commit that changed the dclk clock of the ebc chip to 200
MHz, this is not really required anymore. Therefore, reduce delays to 1
ms.