[rockchip_ebc/panel-simple] Change mode clocks
Now that we control the dclk clock using the dclk_select module
parameter, we can use the display modes to define realistic refresh
rates that refer to a full display refresh, not to one frame.

Add display refresh rates of 20, 25, 30, and 40 Hz.

This should decrease the pressure on the compositors, as they will try
to flip buffers with that frequency.