arm64: pinenote_defconfig: Update net/wireless config
Primary motivation for updating was removing the ``CFG80211_WEXT``
module as Wireless Extensions have been deprecated for over a decade
now, so there really is no reason to enable it for the PineNote.

Furthermore, make the config more in line with Debian's, but don't
include the ones Debian has because they want to support ancient stuff.